chore: Update npm dependencies and remove unnecessary files
This commit updates the npm dependencies in the package.json file and removes unnecessary files from the repository. The ajv package is added as a new dependency, and the @types/react package is updated to the latest version. Additionally, the files "src/assets/js/main.js", "src/firebase.js", and "src/index.js" are excluded from the tsconfig.json file, improving the clarity and maintainability of the repository.